// src/types/trainProps.ts
export interface TrainProps {
  dstn_stn_code: string;
  dstn_stn_name: string;
  from_stn_code: string;
  from_stn_name: string;
  from_time: string;
  running_days: string;
  source_stn_code: string;
  source_stn_name: string;
  to_stn_code: string;
  to_stn_name: string;
  to_time: string;
  train_name: string;
  train_no: string;
  travel_time: string;
}